chore: restructure packages according to MVC pattern
