+++ src/main/java/RacingCar/Main.java
@@ -1,0 +1,10 @@
+package RacingCar;
+
+import RacingCar.controller.RacingCarController;
+
+public class Main {
+    public static void main(String args[]) {
+        RacingCarController controller = new RacingCarController();
+        controller.start();
+    }
+}
